etl/requirements.txt
## For timezones in python
pytz

## For reading yaml files for configuration
PyYAML

## Connect to postgres
psycopg2-binary
SQLAlchemy

## For connecting to github to fetch issues
github3.py

## For development
ipython

## For Airflow
# pytz, see above
pyOpenSSL
ndg-httpsclient
pyasn1
apache-airflow[crypto,celery,postgres,hive,jdbc,mysql,ssh]==1.10.4
redis==3.2 # don't know why 3.2 but it was in the airflow Dockerfile
